# perf(pnp): store inlined pnp data in multiline string literal