**What's the problem this PR addresses?**
<!-- Describe the rationale of your PR. -->
<!-- Link all issues that it closes. (Closes/Resolves #xxxx.) -->

As @merceyz said in https://github.com/yarnpkg/berry/pull/3965#issuecomment-1008178597, [parsing JSON strings is more efficient than parsing JS object literals](https://v8.dev/blog/cost-of-javascript-2019#json).

Closes https://github.com/yarnpkg/berry/pull/3965.

**How did you fix it?**
<!-- A detailed description of your implementation. -->

Made the inlined PnP data get stored in a multiline string literal (where newlines are escaped using `\`) containing JSON.

Why a multiline string literal over a template literal? @merceyz originally looked into using a template literal in https://github.com/yarnpkg/berry/pull/3965#issuecomment-1008178597, but it turned out to be slower than even an object literal. More context: https://github.com/nodejs/node/issues/41448

@merceyz ran some benchmarks on the next.js repo:
```sh
hyperfine -w 5\
  "node -r ./.pnp-object.cjs -e 1"\
  "node -r ./.pnp-string-lf.cjs -e 1"\
  "node -r ./.pnp-data-json.cjs -e 1"\
  "node -r ./.pnp-template.cjs -e 1"
Benchmark 1: node -r ./.pnp-object.cjs -e 1
  Time (mean ± σ):     155.7 ms ±   2.8 ms    [User: 163.1 ms, System: 21.4 ms]
  Range (min … max):   152.3 ms … 161.3 ms    19 runs

Benchmark 2: node -r ./.pnp-string-lf.cjs -e 1
  Time (mean ± σ):     135.1 ms ±   1.8 ms    [User: 135.6 ms, System: 32.0 ms]
  Range (min … max):   132.3 ms … 140.7 ms    21 runs

Benchmark 3: node -r ./.pnp-data-json.cjs -e 1
  Time (mean ± σ):      90.7 ms ±   0.8 ms    [User: 105.0 ms, System: 18.0 ms]
  Range (min … max):    89.3 ms …  92.6 ms    32 runs

Benchmark 4: node -r ./.pnp-template.cjs -e 1
  Time (mean ± σ):     158.0 ms ±   1.2 ms    [User: 160.7 ms, System: 29.2 ms]
  Range (min … max):   154.9 ms … 160.2 ms    18 runs

Summary
  'node -r ./.pnp-data-json.cjs -e 1' ran
    1.49 ± 0.02 times faster than 'node -r ./.pnp-string-lf.cjs -e 1'
    1.72 ± 0.03 times faster than 'node -r ./.pnp-object.cjs -e 1'
    1.74 ± 0.02 times faster than 'node -r ./.pnp-template.cjs -e 1'
```

**Checklist**
<!--- Don't worry if you miss something, chores are automatically tested. -->
<!--- This checklist exists to help you remember doing the chores when you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [X] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).

<!-- See https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check -i` -->
- [X] I have set the packages that need to be released for my changes to be effective.

<!-- The "Testing chores" workflow validates that your PR follows our guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR might be missing. -->
- [X] I will check that all automated PR checks pass before the PR gets reviewed.