@merceyz 2022-01-13
Here are some benchmark results from the various tests I did on the Next.js repo with all examples included
```
hyperfine -w 5\
  "node -r ./.pnp-object.cjs -e 1"\
  "node -r ./.pnp-string-literal.cjs -e 1"\
  "node -r ./.pnp-data-json.cjs -e 1"\
  "node -r ./.pnp-template-literal.cjs -e 1"
Benchmark 1: node -r ./.pnp-object.cjs -e 1
  Time (mean ± σ):     155.7 ms ±   2.8 ms    [User: 163.1 ms, System: 21.4 ms]
  Range (min … max):   152.3 ms … 161.3 ms    19 runs

Benchmark 2: node -r ./.pnp-string-literal.cjs -e 1
  Time (mean ± σ):     135.1 ms ±   1.8 ms    [User: 135.6 ms, System: 32.0 ms]
  Range (min … max):   132.3 ms … 140.7 ms    21 runs

Benchmark 3: node -r ./.pnp-data-json.cjs -e 1
  Time (mean ± σ):      90.7 ms ±   0.8 ms    [User: 105.0 ms, System: 18.0 ms]
  Range (min … max):    89.3 ms …  92.6 ms    32 runs

Benchmark 4: node -r ./.pnp-template-literal.cjs -e 1
  Time (mean ± σ):     158.0 ms ±   1.2 ms    [User: 160.7 ms, System: 29.2 ms]
  Range (min … max):   154.9 ms … 160.2 ms    18 runs

Summary
  'node -r ./.pnp-data-json.cjs -e 1' ran
    1.49 ± 0.02 times faster than 'node -r ./.pnp-string-literal.cjs -e 1'
    1.72 ± 0.03 times faster than 'node -r ./.pnp-object.cjs -e 1'
    1.74 ± 0.02 times faster than 'node -r ./.pnp-template-literal.cjs -e 1'
```